shuttle lighting fix

+++ code/__HELPERS/unsorted.dm
@@ -797,6 +797,9 @@ proc/anim(turf/location as turf,target as mob|obj,a_icon,a_icon_state as text,fl
 //Clear the given turf of anything but living cliets or mobs that have keys
 /proc/clear_turf(var/turf/T)
 	for(var/atom/movable/M in T)
+		//NEVER EVER delete hanger beacons they are fixed to the map
+		if(istype(M, /obj/hanger))
+			continue
 		if(istype(M, /mob))
 			var/mob/mob = M
 			if(mob.client || mob.key)
@@ -936,16 +939,16 @@ proc/anim(turf/location as turf,target as mob|obj,a_icon,a_icon_state as text,fl
 
 
 					for(var/obj/O in T)
-
-						// Reset the shuttle corners
-						if(O.tag == "delete me")
-							X.icon = 'icons/turf/shuttle.dmi'
-							X.icon_state = replacetext(O.icon_state, "_f", "_s") // revert the turf to the old icon_state
-							X.name = "wall"
-							qdel(O) // prevents multiple shuttle corners from stacking
-							continue
-						if(!istype(O,/obj)) continue
-						O.loc = X
+						if(!istype(O, /obj/hanger))
+							// Reset the shuttle corners
+							if(O.tag == "delete me")
+								X.icon = 'icons/turf/shuttle.dmi'
+								X.icon_state = replacetext(O.icon_state, "_f", "_s") // revert the turf to the old icon_state
+								X.name = "wall"
+								qdel(O) // prevents multiple shuttle corners from stacking
+								continue
+							if(!istype(O,/obj)) continue
+							O.loc = X
 
 					for(var/mob/M in T)
 						if(!istype(M,/mob) || istype(M, /mob/aiEye)) continue // If we need to check for more mobs, I'll add a variable
